\name{gets.plm-package}
\alias{gets.plm-package}
\docType{package}
\title{General-to-Specific (GETS) Modelling of linear panel data models of class 'plm'}

\description{
General-to-Specific (GETS) Modelling of linear panel data models of class 'plm', see Pretis, Reade and Sucarrat (2018) <doi:10.18637/jss.v086.i03>.
}

\details{

\tabular{ll}{
Package: \tab gets\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-12-26\cr
License: \tab GPL-2\cr
}

}

\author{
Genaro Sucarrat \url{http://www.sucarrat.net/}\cr\cr
Maintainer: Genaro Sucarrat            
}

\references{
Felix Pretis, James Reade and Genaro Sucarrat (2018): 'Automated General-to-Specific (GETS) Regression Modeling and Indicator Saturation for Outliers and Structural Breaks'. Journal of Statistical Software 86, Number 3, pp. 1-44. \url{https://www.jstatsoft.org/article/view/v086i03}

Genaro Sucarrat (2019): 'User-Specified General-to-Specific and Indicator Saturation Methods'. \url{https://mpra.ub.uni-muenchen.de/96653/}

}

\seealso{
\code{\link{getsFun}}, \code{\link{plm}}
}

\keyword{Statistical Models}
\keyword{Time Series}
\keyword{Econometrics}
\keyword{Climate Econometrics}
\keyword{Financial Econometrics}
